009 Introducing Error Boundaries
this case is demo how to handle Error in the parent component.

we can handle error right now. with Try-Catch

but if we want handle it in the other place, we can create a boundary. and catch the error and display some information elegantly .

componentDidCatch

catch the Error and handle it with  component Did Catch